+++ src/FluxFlow.Components.Timers/TimerErrorCodes.cs
@@ -2,6 +2,7 @@ namespace FluxFlow.Components.Timers;
 
 public static class TimerErrorCodes
 {
+    public const int DebounceFailed = 9040;
     public const int DelayFailed = 9010;
     public const int IntervalFailed = 9000;
     public const int ScheduleFailed = 9020;